# Fix single host mixed throughput